Cleanup old benchmark code
